pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cortex-memory"
version = "0.1.0"
description = "Semantic long-term memory system for humans and AI agents"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Weiming Yin", email = "yinwm@example.com"}
]
keywords = ["memory", "ai", "agent", "semantic-search", "knowledge-base"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pysqlite3-binary>=0.5.0",
    "sqlite-vec>=0.0.1",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/cortex-memory"
Documentation = "https://github.com/YOUR_USERNAME/cortex-memory/docs"
Repository = "https://github.com/YOUR_USERNAME/cortex-memory"
Issues = "https://github.com/YOUR_USERNAME/cortex-memory/issues"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
target-version = "py38"
